requirements.txt
# Database
pymongo==4.6.1
python-dotenv==1.0.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Data Handling
bson==0.5.10

# AI/ML Libraries
pymilvus==2.3.4
sentence-transformers==3.0.1
openai==1.3.7
environs==9.3.5
marshmallow==3.19.0

# Development
python-decouple==3.8

# OAuth
authlib==1.3.0

# Security (already included with Flask)
# werkzeug for password hashing
